requirements_test.txt
coverage==4.4.2
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0


# Additional test requirements go here

# for the demo
django-model-utils
